#14 Create ChannelCollection struct
